fix/use-cache-kv-manager: Moved InvalidateSchemaCacheHandler to a separate module
 • Extracted InvalidateSchemaCacheHandler and associated tests into a new file cache_invalidator.rs
 • Removed async_trait and CacheInvalidator related code from heartbeat.rs
 • Added cache_invalidator module declaration in handler.rs
